requirements/base.txt
Django==1.8.3
psycopg2==2.6.1
pytz==2015.2
django-dotenv==1.3.0
bpython==0.14.2
djangorestframework==3.1.3
djangorestframework-filters==0.3.0
django-filter==0.10.0
django-modeltranslation==0.10
factory-boy==2.5.2
fake-factory==0.5.1
Pillow==2.9.0
exam==0.10.5
rauth==0.7.1
django-model-utils==2.2
geopy==1.10.0
django-rest-auth==0.5.0
celery==3.1.18
django-push-notifications==1.3.1
redis==2.10.5
hiredis==0.2.0
itunes-iap==2.1.0


-e git+git://github.com/comoga/django-constance#egg=django-constance
django-picklefield==0.3.1
